% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_missing.R
\name{clean_missing}
\alias{clean_missing}
\title{Clean Missing Values}
\usage{
clean_missing(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A cleaned data frame
}
\description{
Replaces missing values (NA) with 0.
}
