% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot Summary}
\usage{
plot_summary(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A ggplot object
}
\description{
Creates a simple bar plot of variables in a dataset.
}
