% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleAnalyzer.R
\name{simpleAnalyzer}
\alias{simpleAnalyzer}
\title{SimpleAnalyzer: A simple Shiny App}
\usage{
simpleAnalyzer(x, ...)
}
\arguments{
\item{File, }{The input file in your file system. It should contain at least two columns where the first is the
response and the second is the predictor.}

\item{bins, }{The bins parameter can be controlled via a slider, and it will automatically change the number of
bins of the histogram.}
}
\value{
, Void application
}
\description{
This shiny app will take an input file containing response varialbe in the first column and predictor varialbe
in the second column. A simple histogram is shown for the predicator and a plot where you can select data points
to build a linear regression model.
}
\seealso{
\code{lm}
}
\author{
Mohamed Tleis
}
