% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_stats.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Summarize Statistics}
\usage{
summarize_stats(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Summary output
}
\description{
Returns summary statistics for a dataset.
}
